omg world space sprite rendering!

+++ src/main.cpp
@@ -12,6 +12,8 @@
 #include "res/shader.hpp"
 #include "res/texture.hpp"
 
+#include "lib/glm/gtx/string_cast.hpp"
+
 #include <vector>
 #include <map>
 
@@ -84,27 +86,37 @@ int main(){
 
     //========================================================
     
+    /*
     auto player = global.gameState->ecs->NewObject();
     player->AddComponent<ECTransform>();
     player->AddComponent<ECRigidbody>();
     player->AddComponent<ECSpriteRenderer>();
+    */
 
-    global.resourceLoader->LoadResource<Texture>("tex", "Textures/tex.jpeg");
+    global.resourceLoader->LoadResource<Texture>("tex", "Textures/alignment.png");
     Texture* tex = global.resourceLoader->GetResource<Texture>("tex");
+
+    global.gameState->gameCamera->transform.position = glm::vec3(0,0,10);
+
+    Transform spriteTransform;
+    spriteTransform.scale = glm::vec3(3.0);
 
     while (!global.platform->ShouldClose()){
         global.platform->StartFrame();
-
-        global.renderer->spriteRenderer->DrawSprite(tex, glm::vec2(0, 0), glm::vec2(500,500), 0, glm::vec3(1.0));
         
-        if (global.inputSystem->GetKey(GLFW_KEY_ESCAPE))
-            global.platform->Terminate();
+        if (global.inputSystem->GetKeyDown(GLFW_KEY_ESCAPE)) global.platform->Terminate();
+        if (global.inputSystem->GetKeyDown(GLFW_KEY_F1)) global.platform->SetDisplayMode(Fullscreen);
+        if (global.inputSystem->GetKeyDown(GLFW_KEY_F2)) global.platform->SetDisplayMode(Windowed);     
 
-        if (global.inputSystem->GetKeyDown(GLFW_KEY_F1))
-            global.platform->SetDisplayMode(Fullscreen);
-        if (global.inputSystem->GetKeyDown(GLFW_KEY_F2))
-            global.platform->SetDisplayMode(Windowed);     
-        
+        float camSpeed = 20.0 * global.time->deltaTime;
+        Transform& trans = global.gameState->gameCamera->transform;
+        if (global.inputSystem->GetKey('A')) trans.position -= trans.Right() * camSpeed;
+        if (global.inputSystem->GetKey('D')) trans.position += trans.Right() * camSpeed;
+        if (global.inputSystem->GetKey('W')) trans.position += glm::vec3(0,1,0) * camSpeed;
+        if (global.inputSystem->GetKey('S')) trans.position -= glm::vec3(0,1,0) * camSpeed;
+
+        global.renderer->spriteRenderer->DrawSprite(tex, spriteTransform);
+
         global.platform->EndFrame();
     }
     global.platform->Cleanup();
